pyproject.toml
[project]
name = "prox-challenge"
version = "0.1.0"
description = "Backend agentic architecture for the Prox founding engineer challenge."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.49.0",
  "claude-agent-sdk>=0.1.0",
  "fastapi>=0.115.0",
  "Pillow>=11.1.0",
  "pymupdf>=1.25.0",
  "python-dotenv>=1.0.1",
  "pyttsx3>=2.98",
  "rank-bm25>=0.2.2",
  "sentence-transformers>=3.0.0",
  "uvicorn>=0.34.0",
  "requests"
]

[project.scripts]
prox-backend = "main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
